VoiceVerifier/requirements.txt
django
djangorestframework
django-cors-headers
drf-yasg
numpy